Added testkode and redid portal character errors
