+++ docker/docker-compose.yml
@@ -27,6 +27,16 @@ services:
     networks:
       platform:
 
+  gdoc:
+    image: "fdk/gdoc-import:${TAG}"
+    ports:
+      - "8084:8080"
+    environment:
+     - JAVA_OPTS="-Dspring.profiles.active=docker"
+    networks:
+      platform:
+
+
   fuseki:
     image: "fdk/fuseki:${TAG}"
     volumes:
@@ -80,11 +90,6 @@ services:
 
   elasticsearch:
     image: "fdk/elasticsearch:${TAG}"
-    #command: elasticsearch -Des.cluster.name=fellesdatakatalog
-    environment:
-      - es.cluster.name=fellesdatakatalog
-      - ES_CLUSTER_NAME="fellesdatakatalog"
-      - cluster_name="fellesdatakatalog"
     ports:
       - "9200:9200"
       - "9300:9300"
