+++ harvester/src/main/docker/Dockerfile
@@ -4,7 +4,7 @@ ENV TZ=Europe/Oslo
 RUN ln -snf /usr/share/zoneinfo/$TZ /etc/localtime && echo $TZ > /etc/timezone
 
 VOLUME /tmp
-ADD harvester-?.?.?-SNAPSHOT.jar app.jar
+ADD harvester-?.?.?-SNAPSHOT-exec.jar app.jar
 RUN sh -c 'touch /app.jar'
 
 ENTRYPOINT ["java","-Djava.security.egd=file:/dev/./urandom","-jar","/app.jar"]
