+++ portal/webapp/src/main/resources/templates/search.html
@@ -3,7 +3,7 @@
 </head>
 <body>
 
-<div class="row" th:fragment="search">>
+<div class="row" th:fragment="search">
     <form class="form-group a-form-group site-search col-sm-8 col-sm-offset-2" ation="#" th:action="@{/results}" th:object="${dataitemquery}"  method="post" >
         <div class="input-group input-group-large">
