+++ portal/webapp/src/main/resources/templates/theme.html
@@ -35,7 +35,7 @@
 
                 </h4>
             </div>
-        </th:block>z
+        </th:block>
     </div>
 
     <div lang="en" th:replace="footer :: footer" />
